Give the stub a canvas, and let the checks find their own inputs
Five harnesses each carried their own copy of the minimal DOM they evaluate a
page in, and all five had getContext() return null. That was harmless until the
viewer gained its animated colour-cycling footer, which asks for a 2D context
while the script body is still loading: three checks then failed at load, and
the two that would also have failed were skipping instead, because the delvmod
checkout they need was never there. One stub now, in dom_stub.mjs, with a real
context -- and viewer_smoke.mjs shares its canvas so the two cannot drift apart
again.

Loading successfully turned out to have a second effect: the footer's timers
now actually register, and a repeating timer holds Node's event loop open, so a
harness that had printed its result would sit there forever. Intervals are
unref'd for that. One-shots are left alone, since a one-shot is often the thing
being waited for.

The inputs find themselves now. check_all.mjs reads the committed archives out
of res/ rather than requiring a sources/ directory to be symlinked into place
first -- a step easy to forget, and forgetting it reported most of the suite as
"skip", which reads like a clean run. delvmod comes in as a submodule at
reference/delvmod so the two checks that read it can actually run; $DELVMOD and
$INFINITE_MAC point at copies kept elsewhere.

The delvmod cross-check passes on first run: all 1,558 resources agree on
decryption, and every table matches.

14 checks run, 0 failed, 1 skipped, the skip being the infinite-mac checkout
that is gitignored on purpose. It was 12 run, 3 failed, 3 skipped.
