+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 repomix_output.md
 cythera/infinite-mac
+cythera/sources
