+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "cythera/reference/delvmod"]
+	path = cythera/reference/delvmod
+	url = https://github.com/e-z-g/delvmod
